requirements.txt
ollama
pyfiglet
ddgs
requests
beautifulsoup4
readability-lxml